requirements.txt
flask
flask-cors
werkzeug
numpy
opencv-python
pillow
scikit-learn
matplotlib
pandas